Update statistics and instruction chapter
This change includes a bugfix in the example code, and a short section
on vector packet processing (including a glossary update), in relation
to statistics. It also adds information on what compiler flags are
used in the counter benchmarks.

The introduction chapter also receives a small update, fixing a
reference to 5G RAN CU, which should be CU CP.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>