Describe the standard threading model
The change introduces a section the on typical network application
architecture (the "standard model"), primarily in regards to
threading, in the threading chapter.

Included is also the result of a number of different benchmarks, in an
attempt to quantify the cost of the standard model, and highlight why
it in many cases cannot be used in data plane applications.

The glossary receives some significant updates as well.

The threading chapter still very much a draft version, with a lot of
holes in the form of unwritten sections, is missing diagrams, and
lacks in quality.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>